outlook_project/backend/requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
python-dotenv==0.19.2
